django_server/requirements.txt
asgiref==3.2.10
astroid==2.4.2
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-filter==2.3.0
django-oauth-toolkit==1.3.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
google-auth==1.19.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
oauth2-provider==0.0
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.5.3
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
wrapt==1.12.1
